% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roboost-pls-da.R
\name{predict_roboost_plsr}
\alias{predict_roboost_plsr}
\title{Prediction with RoBoost-PLSR}
\usage{
predict_roboost_plsr(mod, Xu)
}
\arguments{
\item{mod}{: RoBoost-PLSR model}

\item{Xu}{: New dataset}
}
\value{

}
\description{
Prediction with RoBoost-PLSR
}
\examples{


}
