% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{r2.t}
\alias{r2.t}
\title{rmse trimmed}
\usage{
r2.t(x, y, W, z)
}
\arguments{
\item{x}{residuals}

\item{y}{Y responses}

\item{W}{weigths}

\item{z}{trimmed threshold}
}
\value{

}
\description{
rmse trimmed
}
\examples{


}
