% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roboost-pls-da.R
\name{roboost_bcoef}
\alias{roboost_bcoef}
\title{bcoef for RoBoost-PLSR}
\usage{
roboost_bcoef(mod, ncomp)
}
\arguments{
\item{mod}{: RoBoost-PLSR model}

\item{ncomp}{: the number of lv to estimate the model}
}
\value{

}
\description{
bcoef for RoBoost-PLSR
}
\examples{


}
