% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roboost-pls-da.R
\name{roboost_plsr}
\alias{roboost_plsr}
\title{RoBoost-PLSR : Robust method for partial least squares regression}
\usage{
roboost_plsr(
  X,
  Y,
  ncomp,
  niter = 50,
  gamma = gamma,
  beta = beta,
  alpha = alpha,
  th = 1 - 10^-12
)
}
\arguments{
\item{X}{Explanatory variables}

\item{Y}{Explained Variables}

\item{ncomp}{Number of latent variables}

\item{niter}{Number of maximal iterations}

\item{gamma}{parameters for leverage point}

\item{beta}{parameters for Y-residuals}

\item{alpha}{parameters for X-residuals}
}
\value{

}
\description{
RoBoost-PLSR : Robust method for partial least squares regression
}
\examples{

n <- 10
p <- 6
set.seed(1)
X <- matrix(rnorm(n * p, mean = 10), ncol = p)
y1 <- 100 * rnorm(n)
y2 <- 100 * rnorm(n)
Y <- cbind(y1, y2)
set.seed(NULL)
Xr <- X[1:8, ] ; Yr <- Y[1:8, ]
Xu <- X[9:10, ] ; Yu <- Y[9:10, ]
library(roboost)
ncomp = 3
alpha = Inf
beta  = Inf
gamma = Inf
mod = roboost_plsr(X = Xr,Y = Yr ,ncomp,gamma =gamma,beta=beta,alpha=alpha)
pred = predict_roboost_plsr(mod$fm,Xu)
}
