% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{segmkf}
\alias{segmkf}
\title{Segments for cross-validation}
\usage{
segmkf(
  n,
  y = NULL,
  K = 5,
  type = c("random", "consecutive", "interleaved"),
  nrep = 1,
  seed = NULL
)
}
\arguments{
\item{n}{: number of samples}

\item{y}{: A vector of length n defining the blocks.}

\item{K}{: number of fold}

\item{rep}{: The number of replications of the repeated CV. Default to nrep = 1.}

\item{For}{segmkf. The type K-fold CV. Possible values are "random" (default), "consecutive" and "interleaved".}
}
\value{

}
\description{
Segments for cross-validation
}
\examples{


}
